+++ packages/express/src/createAuth.ts
@@ -2,7 +2,7 @@ import { createAuth as createBasicAuth, type AuthConfig } from "@aura-stack/auth
 import { withAuth } from "@/lib/with-auth.ts"
 import { toExpressHandler } from "@/lib/handler.ts"
 import type { Request, Response } from "express"
-import type { EditableShape, ShapeToObject, UserShape } from "@aura-stack/auth/identity"
+import type { EditableShape, ZodShapeToObject, UserShape } from "@aura-stack/auth/identity"
 
 export const createAuth = <Identity extends EditableShape<UserShape>>(config: AuthConfig<Identity>) => {
     const auth = createBasicAuth<Identity>(config)
@@ -16,6 +16,6 @@ export const createAuth = <Identity extends EditableShape<UserShape>>(config: Au
         /**
          * Middleware that retrieves the session and attaches it to `res.locals.session`.
          */
-        withAuth: withAuth<ShapeToObject<Identity>>(auth),
+        withAuth: withAuth<ZodShapeToObject<Identity>>(auth),
     }
 }